Import Code
This imports the UI code from https://github.com/allenai/llmx
with minimal changes.

The main derivation involved mutating the `LLMX_API_URL` variable
so that it's always required, and defaults to something that works
in development.

I also added `proxy/local.conf.template` for local environment and
updated instructions for port-forwarding to the local API. We'll
make this better soon.