Don't auto expand first item
This changes the behavior of the thread list(s) so that the first
accordion item isn't expanded by default, unless a new message is
being streamed to the client.

To accomplish this I lifted the bit that tracks the expanded thread
into the global context, and added method that changes it.

I also executed a small refactor, changing the `threadKey` property to
`threadID`, as I'm not sure what a `threadKey` is.

https://github.com/allenai/olmo-ui/issues/45