Use cookies (#53)
I recently tweaked the API to set HTTP cookies when hitting /v3/whoami.
This means the UI can stop ferrying around the token and rely on the
browser to do so instead. This has many benefits. It's more secure,
allows users to hit the API directly in a browser after logging in once,
and it makes the UI code simpler.

After this merges I'm going to update the API to stop returning the
token in the JSON response, as it won't be required for the client
anymore.