Remove Auth (#75)
This removes auth, so that invite URLs work. To do this I reworked the
top-level loading experience and what we display when the user isn't
logged in. We want to "leak" as little as possible, so I elected to hide
the header, logo and navigation elements in that scenario.

Here's the new loading experience:

<img width="1109" alt="Screenshot 2023-09-26 at 10 06 25 AM"
src="https://github.com/allenai/olmo-ui/assets/801451/f4f70554-b631-4b80-9b64-4478e1707f01">

And what we display when the user isn't logged out:

<img width="1109" alt="Screenshot 2023-09-26 at 10 06 07 AM"
src="https://github.com/allenai/olmo-ui/assets/801451/1f86c86c-3cd6-44e4-8553-7ce2260cb223">

We can polish this later. We need to do something quick now for the SAB
demo this friday.