Fix auth (#79)
This is the UI portion of https://github.com/allenai/olmo-api/pull/31.

After that change lands the API will return 401s to `/v3/whoami` when
the client isn't authenticated. When that occurs the UI will take the
user to `https://olmo-api.allen.ai/v3/login/skiff`, which will bounce
the user through the usual authentication process.

This fixes the issue where users see the "Unauthorized" message now when
not logged in, w/o a way to resolve it. And it means invite tokens will
work as expected, as clients w/ valid tokens (in their HTTP cookies)
won't be bounced through Skiff Login.