Better first load experience
@aimichal noticed that the "Unauthorized" message shows while loading
the UI while not logged in. This fixes that.

The issue was caused by the fact that the "stores" are initially not
set, which means `userInfo.loading` is `false`.

This fixes that by relaying on local state instead which will be
toggled appropriately. This means the loading indicator displays
instead, as desired.