Handle search errors
This modifies the search page so that API errors are handled
gracefully by displaying the error to the end user. This depends
on https://github.com/allenai/olmo-api/pull/36.

To do this I did some factoring to reuse code in `AppContext` for
"unpacking" common errors. Again this is a small step forward that'll
get better as we make time to cleanup API interactions / how we persist
state on the client.