Handle search errors (#83)
This modifies the search page so that API errors are handled gracefully
by displaying the error to the end user. This depends on
https://github.com/allenai/olmo-api/pull/36.

To do this I did some factoring to reuse code in `AppContext` for
"unpacking" common errors. Again this is a small step forward that'll
get better as we make time to cleanup API interactions / how we persist
state on the client.

I also fixed some scenarios where we weren't properly updating/resetting
state. For instance, if the user navigates from `/search` to
`/search?query=Seattle`, then hits back to return to `/search` we'll now
properly clear the results and display the placeholder.

<img width="1105" alt="Screenshot 2023-09-27 at 10 06 37 AM"
src="https://github.com/allenai/olmo-ui/assets/801451/224132c0-92f2-40c1-af21-ffb98e398e6c">