Flesh out an API client
This fleshes out the foundations of an API client by replacing the
code on the search pages with code that uses a client.

The idea is to lift the API logic into a library that nicely
encapsulates things. Then localize state management into the
components themselves.

I find the code a bit easier to follow and believe it affords some
flexibility the current pattern doesn't. Embedding API interactions in
the store feels like a coupling we should avoid.

I wanted to get this pattern in place/proposed, but don't plan on
pushing on it further. I figured we could merge this to set the
precedent but that I'd let others extend this pattern across the
application.

https://github.com/allenai/olmo-ui/issues/107