Flesh out an API client (#109)
This fleshes out the foundations of an API client by replacing the code
on the search pages with code that uses a client.

The idea is to lift the API logic into a library that nicely
encapsulates things. Then localize state management into the components
themselves.

There's some duplication of code in `AppContext`. I figured we'd replace
that as things get refactored, rather than replacing it now (which is
riskier). I toyed w/ the idea of using local `zustand` driven stores,
but found a interface + `setState()` to be enough and preferable for
it's simplicity.

I wanted to get this pattern in place/proposed, but don't plan on
pushing on it further. I figured we could merge this to set the
precedent but that I'd let others extend this pattern across the
application.

https://github.com/allenai/olmo-ui/issues/107

---------

Co-authored-by: jonathan m borchardt <jonathanb@allenai.org>