Fix #146; Split stream on non-escaped newlines
This fixes #146. The current mechanism splits on any newlines, which
means if the generation includes a newline the generation stops when
the UI fails to try to parse invalid JSON (and throws an exception).
The UI should handle this scenario better, but I'll queue that as a
separate PR to keep this easy to review.

The patch modifies the split so that we only split on `\n` characters
that aren't escaped (i.e. `\\n`). JSON requires that newlines be
serialized, so this should be reliably. To do this we use a regular
expression with a negative lookbehind.

I also refactored the stream processing a bit. The `done` bit
will only be set to `true` when `value` is null. This makes the
code a little easier to read, I think.