Fix bug that sometimes when hover over the "..." menu, we can only click when the cursor is not over the dots. (#154)
## Description

Ref: https://github.com/allenai/olmo-ui/issues/144
Report by Michael and i saw u was facing the same problem during the
OLMO-ui demo. This PR addresses this

## Reviewer Instructions

It turns out the `<IconContainer />` container pointer event interfere
with the pointer-event of the mui Button. Since `<IconContainer />`
doesnt do anything beside act as a wrapper and has style, I turned off
the pointer event of it and it allows the button to be clickable again.

## Testing Plan

1. Verify UI works.
2. If CI passes we are good.

## Output / Screenshots

### Before



https://github.com/allenai/olmo-ui/assets/84343285/6af14cc4-e9d3-45b0-a0fc-26a1286cdb10


### After



https://github.com/allenai/olmo-ui/assets/84343285/9d2c0c90-a411-41f4-a7e6-1a98da103ec2


### A11y

N/A