Update readme instruction on how to set up verify commit gpg way (#155)
I ran into issue of try to commit and merge the branch in Olmo-ui. I
went down the rabbit hole of how to set up gpg and make sure commit is
verify so I want to document my step incase if any one runs into this
issue. Feel free to give me feedback if the instruction is clear or not
@arnavic