Add a checkbox to allow users to view private queries. the checkbox position is under My Queries (#169)
## Description

Ref: https://github.com/allenai/olmo-ui/issues/168

This PR add a toggle with label Private Queries under My Queries (it
won't show in All Queries). If toggle is on we only show private query
off we will show regular query.

## Reviewer Instructions

One thing to know since the private query can only be applied for
current pagination and it cant be apply for the whole full blown. I
think if we want to get only private queries when toggle the API may
need to return the whole full sweep of them and then we can just render
instead of filtering out based on the current page.

## Testing Plan

Verify it works in UI.

## Output / Screenshots


https://github.com/allenai/olmo-ui/assets/84343285/676eda00-0367-459a-8438-a011a77b1fc5



### A11y

N/A