Setting up toggle for basemodel while we are still in development (#195)
## Description

Setting up toggle to disable base model by default in the dropdown list
of models while we are still in development. This assumes we have a
field on Model that is model_type (per discussions with MichaelW).
@mewil I've made model_type optional for now so everything still works
as expected but I will switch this out once you make the change.