remove datachip UI code (#223)
## Description

Closes #220

This PR removes code related to datachips in the UI. There is another
branch named
[`withdatachips`](https://github.com/allenai/olmo-ui/tree/withdatachips)
that contains the code as a sort of save point.

Since we're not using datachips right now and don't have plans to in the
near future we thought it would be best to remove it from the repo for
now. The code is still in place on the API.

## Reviewer Instructions

Are there any uses of datachips I've missed? Is there anything I'm
removing that we may want to keep?

## Testing Plan

Ran the app locally and ensured no datachips showed up and the UI still
functions correctly.