Add MSW to New Query tests (#234)
## Description

This PR adds MSW (a network request mocking framework) to olmo-ui to
make it easier to fake requests, including the message stream.
It is currently only enabled in one test. It's only enabled in the
browser if the `ENABLE_MOCKING` flag is set to `true`.

I replaced mock fetch setup in NewQuery tests in favor of MSW. It'll be
nice to keep it all in one place. If we want to enable it for all tests
I can set that up as well! LMK!