move fetching logic from AppContext into respective Clients (#232)
## Description

Closes #216 

I followed the existing pattern on `PromptTemplateClient` to extract API
logic out of the store. I kept the Clients in the same file as their
types, though. If we want to change that it'd be easy to make them a
folder and keep them grouped that way!

## Reviewer Instructions

what do you think of using the Client pattern? should we extract more
from the AppContext?

## Testing Plan

Tested all functions I could find on the UI locally. Haven't introduced
unit tests since the calls are all _fairly_ straightforward (except
sendMessage and that's hard to test!)


Example of a prompt working: 

![apiclientrefactor](https://github.com/allenai/olmo-ui/assets/2118649/1878bcf7-a547-4934-9481-119124394d2c)