Adding Dolma back into OLMo-UI (#241)
## Description

This PR adds Dolma back into the olmo-ui repo. With this PR, you can hit
olmo.allen.ai/dolma, and it will pull up what is currently on the Home
page at dolma.allen.ai. We have not yet made stylistic changes or
adjustments to the content.

## Reviewer Instructions

The Dolma .tsx files will probably change quite a bit with latest
designs, so try not to review those with a lot of scrutiny; instead,
would love feedback on how we're integrating this to be hosted in
conjunction with the olmo ui and api.

## Testing Plan

Local testing of various Dolma charts and searches.

## Output / Screenshots

### After


![image](https://github.com/allenai/olmo-ui/assets/6945628/b28c066b-2fa8-4c0d-90ad-1bf29ef7b181)


![image](https://github.com/allenai/olmo-ui/assets/6945628/67418a25-03dd-497f-8ea7-e9fee52df1a2)

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>