add env feature flag for the v0 UI (#253)
This PR adds a feature flag toggle to enable the new v0 UI. Chose to use
an env feature flag over stuff in the FeatureToggleContext because
react-router wants things to be defined outside of the component tree.

With this, we can add things like navigation to the layout inside of the
v0 routes and it won't affect the pre-v0 routes.

if in an environment where the new UI isn't enabled by default, you can
visit http://localhost:8080/?isUIRefreshEnabled=true to enable it

Tested that toggling the env on/off worked as intended