SearchStore -> Zustand  (#263)
Migrates over SearchStore to use the Zustand pattern, #251 . 
I validated that all is working fine with searches that have responses
as well as searches that have no responses! Screenshot below.
<img width="1243" alt="Screen Shot 2024-04-01 at 11 56 55 AM"
src="https://github.com/allenai/olmo-ui/assets/6945628/6d299026-5090-44a0-ba9b-b7642c7665d3">


Note: This PR also changes some text colors for the Dolma Search
results, I did this because we have to record a video this week for NSF
with the old UI and the existing search response colors were not
sticking out well on the grey/black background. This will change as we
work on Dolma reskinning over the next week-ish.