New Thread page start (#275)
## Description

I'm submitting this as a PR early because it has a ton of work I want to
have in main. This contains the work I've done so far on the updated
thread UI. It has some updates to nav and layout styling along with the
skeleton of the thread page. It's missing thread submission and
updating, responsive layout for the thread buttons at the top, the
buttons under each thread, and a lot of interactivity.

## Reviewer Instructions
Don't worry about functionality right now. Do the components look good?
How does the layout look?


https://github.com/allenai/olmo-ui/assets/2118649/bf130bbd-3b42-4389-a3de-e9e4171a5d8c

Desktopish:

![image](https://github.com/allenai/olmo-ui/assets/2118649/d779ea2f-e77d-404a-b0e9-e5fde3a823ac)

![image](https://github.com/allenai/olmo-ui/assets/2118649/c23ea87e-9e53-45bb-83f5-25e50b8b075a)

Mobile: (the buttons and container for them are a WIP!)

![image](https://github.com/allenai/olmo-ui/assets/2118649/aa89662b-f115-415a-b4f6-43cfa286e166)

![image](https://github.com/allenai/olmo-ui/assets/2118649/d553c7e8-c9ab-4fa3-a4ff-4c599be8b443)