Add functionality to send a new message (#283)
## Description

This PR does two big things:

refactors threads and message sending
adds the ability to create threads and respond to them

There are details about changes in comments on the PR

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Unit, e2e tests all passing.

Manual testing works well in old and new UIs

## Output / Screenshots

### Old UI:

https://github.com/allenai/olmo-ui/assets/2118649/d77c6c13-5656-4a16-a947-973d4fbb8667

### New UI:

https://github.com/allenai/olmo-ui/assets/2118649/ed5896f8-b631-4be1-a080-32fe35fd9b6e

### A11y
Keyboard nav works as expected. We may need to add aria attributes to
announce that the thread is changing.