Global analytics client (#319)
I've confirmed that the analytics client works to log both olmo and
dolma events. This PR makes a global client instance that gets imported
into specific playground pages and components. It also adds a tracking
event for Follow Up prompts in OLMo.
I'm going to check in with PM and Design about what else they
specifically want to log, and create a ticket for that.
Next steps would be to: 
1) Integrate with Plausible 
2) Implement more event tracking based on PM and Design's responses 

Logs can be seen in the Preview tab here:
https://console.cloud.google.com/bigquery?project=ai2-reviz&ws=!1m5!1m4!4m3!1sai2-reviz!2sdolma!3sevent