Add info button to stop words parameter (#392)
## Description

Closes #365 

This PR adds an info button to the Stop Words input in the parameters
drawer. It also pulls the info button and tooltip into a separate
component to make it easier to reuse.

## Testing Plan

Tested locally. Testing dialogs with RTL is a pain :(

## Output / Screenshots

### Before


![image](https://github.com/allenai/olmo-ui/assets/2118649/dd8a6b5f-aa17-484a-a85a-e56f9df8535e)

### After


https://github.com/allenai/olmo-ui/assets/2118649/1fc69642-70fa-4ca7-a2fc-1643a51d46b8

### A11y

We now follow [best practices for
tooltips](https://www.w3.org/WAI/ARIA/apg/patterns/tooltip/) by allowing
users to close them with Escape. Using the semantic `label` element for
the input gives us the same compliance as the MUI input.

---------

Co-authored-by: Huy Tran <84343285+huytr1995@users.noreply.github.com>