Make the easy parts of the FAQ page with fake content (#402)
## Description

Addresses #401 

This PR adds a basic FAQs page, including an `FAQCategory` component and
an `FAQ` component. It allows for linking to the category or to an
individual FAQ.

It does not include categories or a search function.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally.

## Output / Screenshots


https://github.com/allenai/olmo-ui/assets/2118649/0e771b6a-dc07-4990-b958-26c3ef237ec3

### A11y

Followed a11y requirements from w3c for accordions