Add Delete Thread Confirmation Dialog & Snackbar (#393)
## Description

Ref: https://github.com/allenai/olmo-ui/issues/381
Closes: https://github.com/allenai/olmo-ui/issues/381
Figma: 
Mobile:

https://www.figma.com/file/HzdNG5hmhXtRYV5f7Ed7G3/OLMo-Platform-V0-(2024)?type=design&node-id=1117-5419&mode=design
Desktop:

https://www.figma.com/file/HzdNG5hmhXtRYV5f7Ed7G3/OLMo-Platform-V0-(2024)?type=design&node-id=1117-7721&mode=design&t=ytSHOACjVOcVPiIJ-0

Discussed with @mtblanton that Cassidy has two different design for both
mobile & desktop. But one is confirm dialogue, the other is like a
drawer so we agree to let make it a confirmation dialogue like the
desktop version.

## Testing Plan

Verify UI works. If CI passes too we are good.

## Output / Screenshots



https://github.com/allenai/olmo-ui/assets/84343285/a8dd4b5b-276d-47b0-9b75-abcc6cee84aa


### A11y

N/A