Add search to FAQs (#413)
## Description

Closes #410 

This PR adds search/filtering functionality to the FAQs page.

The performance on filtering seems a little slow? I tinkered with it a
little and didn't have an easy solution. Not sure if the slowness is
from React re-rendering everything or the filtering itself.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally. Testing with React-Router integration is annoying!

## Output / Screenshots
No search:

![image](https://github.com/allenai/olmo-ui/assets/2118649/8e55fab9-922e-48c4-ad10-8fe642dbfdc6)


Search with results from two categories:

![image](https://github.com/allenai/olmo-ui/assets/2118649/8eba468e-ae0b-4e96-ac7e-33cdf8dbc95c)


Search with no results:

![image](https://github.com/allenai/olmo-ui/assets/2118649/252c8e1a-65f7-44cd-b4e7-d7bf57c90bbb)

Dolma search w no results:

![image](https://github.com/allenai/olmo-ui/assets/2118649/dcd1a9f6-d15d-42c0-8da9-92d793e9052c)