Fix submit issue on empty thread (#443)
## Description
Closes: https://github.com/allenai/olmo-ui/issues/448
Reported by Eric that when he hit the submit button without any content
in the text box. It gives him an error for unable to process the
request. This PR addresses this issue.

## Reviewer Instructions

This is because empty content still being able to send to the back end.
I added a check to make sure content is not empty before submit.

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

![image](https://github.com/allenai/olmo-ui/assets/84343285/c9594ffe-2cb3-456a-93b8-f5f5e568447c)

### After



https://github.com/allenai/olmo-ui/assets/84343285/50daea6a-b2df-4e29-96cb-5ac8d1c25c88



### A11y

N/A