(UI) Block non-creator from responding to a thread (#452)
## Description

Closes: https://github.com/allenai/olmo-ui/issues/446
To prepare for public release we wouldnt want user can just prompt on
other people thread. This PR removes the prompt box and submit button if
user is viewing a thread that not created by themselves.

## Testing Plan

Verify by giving local my threadId of my allenai email on prod to see if
both textarea & submit button not showing. If CI passes we are good too.

## Output / Screenshots


![image](https://github.com/allenai/olmo-ui/assets/84343285/456d952a-3333-430a-b3a5-69600e27fbd1)


### A11y

N/A