Update to use Dolma 1.7 (#501)
## Description

closes #470

Upgrades to using the Dolma 1.7 index. Keeping the 1.5 configs around
just in case we need to switch back quickly.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally

## Output / Screenshots

![image](https://github.com/allenai/olmo-ui/assets/2118649/08cf78b6-0567-4288-b850-358b0a96a1be)