Fix history is not reflecting as far back as it used to (#527)
## Description

Closes: https://github.com/allenai/olmo-ui/issues/520

## Reviewer Instructions

In history drawer we have a logic to determine whether or not we can
continue to do infinite scrolling. But we forgot to store the meta data
total of messageList in it so we can validate the logic. Adding this in
ThreadSlice solve the problem.

## Testing Plan

Verify UI works by continuous scrolling. If CI Passes we are good.

## Output / Screenshots


https://github.com/allenai/olmo-ui/assets/84343285/282109aa-e1bc-4890-9458-8c1faab26261

### A11y

N/A