Update the sources chart for mobile (#534)
## Description

Closes: https://github.com/allenai/olmo-ui/issues/512
Figma:
https://www.figma.com/design/HzdNG5hmhXtRYV5f7Ed7G3/AI2-Playground-(2024)?node-id=2422-6480&t=rhpeoaQ4hYsreeJp-0

## Reviewer Instructions

We decided to implement a bar chart but display in a horizontal from top
to bottom. This help to make it looks better. Im also added the property
to make it can be tab by the graph or each bar itself.

## Testing Plan

Verify UI works. If CI passes too we are good.

## Output / Screenshots


https://github.com/allenai/olmo-ui/assets/84343285/4c2e3cd7-c281-4688-a038-0503514f015b

### A11y

N/A