[OEUI-66] highlight spans from the selected document (#544)
## Description

Addresses part of OEUI-66

This PR adds highlighting for selected spans (but not currently preview
(hover) spans)

This solution uses
[remark-directive](https://github.com/remarkjs/remark-directive) and
[remark-directive-rehype](https://github.com/IGassmann/remark-directive-rehype)
to highlight the spans and show them in the rendered markdown.

It currently has issues with spans that contain newlines and spans
inside of code blocks. There may be other issues that I haven't run
into!

## Testing Plan

Tested locally.

## Output / Screenshots


https://github.com/user-attachments/assets/592f374e-4e26-4285-9b6c-78950ba21bdf

### A11y

There's no announcements for highlights changing, though I'm not sure we
need them? Do we need to announce what's changed if it's a
user-initiated change? I guess we probably need good labels on the
cards?