[OEUI-66] Highlight spans from documents that are hovered or focused (#545)
## Description

Closes OEUI-66

This PR builds off of the selected span highlights PR and adds
highlighting for documents that are hovered or focused. If a span is
duplicated in the selected or previewed documents we only show the
selected highlight.

## Testing Plan

Tested locally.

## Output / Screenshots


https://github.com/user-attachments/assets/20b47347-bfc4-4bcc-9dcd-3222aafa766c


### A11y

Same as in the other PR, not sure what announcements we need to do. We
do support keyboard and mouse nav though!