[OEUI-102] Set up making spans clickable, add a feature flag for it (#552)
## Description

Addresses OEUI-102

This PR sets up other PRs to allow clicking spans to show different
things in the attribution drawer. Clicking a span currently doesn't do
anything visually, but it does set some state in the store.

## Testing Plan

Tested locally with live and mock data.

## Output / Screenshots
The only visual here is the action in the redux devtools!


https://github.com/user-attachments/assets/f228d265-cc80-4a0e-9d56-ecacc986488c

### A11y

The clickable spans are focusable buttons and have a label describing
what they do. We may want to change the focus styling later, but that's
not for this PR!