[OEUI-102] Highlight all spans (#553)
## Description

Closes [OEUI-102]

This PR makes all spans in a response be highlighted by default. It
disables the functionality to select or preview a document and have the
spans associated with that document be highlighted.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Local testing, with and without mocks.

## Output / Screenshots


![image](https://github.com/user-attachments/assets/6618eb12-8e07-4f00-8917-6cca73c7b85c)

### A11y

I've tried to ensure buttons are still focusable and have a good label.

[OEUI-102]:
https://allenai.atlassian.net/browse/OEUI-102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Huy Tran <huytr1995@users.noreply.github.com>