[OEUI-108] Show documents that include the selected span in the drawer (#555)
## Description

Addresses OEUI-108

This PR makes it so that you can click on a highlighted span and see
only documents that that span is in. This doesn't add the ability to
show/hide all spans or de-select a span.

There's some reset logic we didn't need to add until now in the loaders.

## Testing Plan

tested with mocks and live data

## Output / Screenshots


https://github.com/user-attachments/assets/f2722621-44fc-49f1-a402-0bfecbf038ed

### A11y

We need to make sure we're announcing what's needed if the filtering
changes in the drawer.