[OEUI-112] Escape markdown when adding attribution highlights (#560)
## Description

Closes OEUI-112

This PR sets up regex escaping for span searches so we can highlight
spans with regex-reserved characters in them.

It also refactors our test setup to be able to inject whatever values we
want into it. It uses a custom FeatureToggleProvider and the
ThreadDisplay test file has a custom Context that lets us inject
arbitrary values. I'd like to pull this out to a reusable thing in the
future but I'm not sure how it works with the mocks and such.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally, added a unit test to check various regex character
combos

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/1a1fc1d3-ecad-46de-a218-2cad0e58a1e3)

### After


![image](https://github.com/user-attachments/assets/6fd24c49-7614-4c20-a4fc-c6066c67156b)

### A11y

No a11y changes