Order documents by the position of the first span inside of it (#563)
## Description

Closes OEUI-116

## Reviewer Instructions

The span has been returned in their appearance, so In `AttributionSlice`
I created an array to store all the documentId of each span. And in
`message-attribution-documents-selector.ts` I reorder the documents list
based on the their documentId list and it will display correctly on the
front end based on the position of the first span inside of it.

## Testing Plan

Verify UI works by testing a couple thread and check if the first
document match the first span showing in the text. If CI passes we are
good.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/eed04a8d-64ec-443a-a011-66a95fdc50ad


### After



https://github.com/user-attachments/assets/fb0ad8db-f3c4-4a64-8841-ee6cde20a890


### A11y

N/A