Fix ordered list styling in chat responses (#567)
## Description

Closes: https://allenai.atlassian.net/browse/OEUI-46

## Reviewer Instructions

We added a `white-space: preserve` previously in `ChatMessage.tsx` which
cause this. From what i understand `rehype-raw` in conjunction with
`white-space: preserve` could cause this because for `white-space:
preserve` can cause text to not wrap correctly and may create extra
space in layout. When combined with raw html prasing (viva rehype-raw)
it can cause these issue, especially if HTML elements have their own
white space or styling. I decided to get rid of it and test with other
thread has order list or bullet point to see if its still working.

## Testing Plan

Verify both the link that being sent still work and no line break or
weird styling for order display. If CI passes we are good.

## Output / Screenshots

### Before


![image](https://github.com/user-attachments/assets/c4b18358-8859-4820-b423-c17fa9b73591)

### After


![image](https://github.com/user-attachments/assets/62cb5eb6-51bf-49f0-8ae9-39e88f3983f4)


### Other testing 

![image](https://github.com/user-attachments/assets/dd8aa0c1-5ef5-4c63-8941-b02db721d2e1)

![image](https://github.com/user-attachments/assets/5ca77ce0-81f0-4a2e-ae5c-31f4fed7e605)


### A11y

N/A