[OEUI-83] Prevent attribution-highlight directives from showing up in code blocks while still allowing users to select it to show spans (#566)
## Description

Closes OEUI-83

This PR makes it so that we don't have attribution highlight directive
code showing up in code blocks. It also makes it so that clicking on a
code block will be similar to clicking on a highlighted span. It'll show
documents associated with any of the spans inside the code block.

## Testing Plan

banged my head against markdown for a while

## Output / Screenshots


https://github.com/user-attachments/assets/fa56d7d2-5868-4b2e-a91f-f753393e6415


### A11y

The button is focusable and has good text