Fix spn selection doesnt update documents after selecting specific spans (#572)
## Description

Closes: https://allenai.atlassian.net/browse/OEUI-134

## Reviewer Instructions

I noticed in my dev console that there is a lot of duplication like
snapshot below. So I decided to do filter out all the duplicate in the
documents before we return them to render and it solves the problem.

![image](https://github.com/user-attachments/assets/4e446815-f459-4f9c-a579-f7dc135f32cb)

## Testing Plan

Verify UI works by continue clicking different span and make sure they
update accordingly. If CI works too we are good.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/44dd4cae-960f-457b-a223-e3dbc0172759

### After


https://github.com/user-attachments/assets/e5f4d101-1e09-4e65-90c8-69e487c81de3

### A11y

N/A