[OEUI-169] rework the router to make the 404 page not be in auth (#583)
## Description

Addresses #581, OEUI-169

This PR reworks our routes so that the 404 catchall route isn't under
auth. Everything else that needs to be is still under it.

We think _something_ in the JS code in the Auth0 change caused issues so
I've been trying to divine what would cause that. I've had some luck in
other test branches with this so I figured we could try it out on dev.

## Testing Plan

Tested locally with 404 and normal pages