[OEUI-170] fix the T&C modal crashing after finishing (#587)
## Description

Closes OEUI-170

The application was crashing after accepting t&cs because we were out of
bounds of the sections array. This changes the form submission behavior
so that it doesn't move the `activeStep` up one if we're on the last
one.

## Testing Plan

Tested locally, added a `showTermsAndConditions()` function to root so
we can test after we've already accepted, added a unit test.