[OEUI-123] Show a message when we get an `inappropriate_prompt` response from the API (#577)
## Description

Closes OEUI-123

This PR adds handling for `inappropriate_prompt` validation responses
from the server. It also changes the logic for showing the "stop" button
so it shows as soon as we're loading a request. Wildguard adds a little
time before the first response comes back so this lets us show the stop
button as soon as we're loading.

## Testing Plan

Tested locally. Need to add an e2e or unit test.

## Output / Screenshots


https://github.com/user-attachments/assets/47417b09-35d9-4b0e-9f7f-065577be7f09

## a11y

Relies on MUI's error message implementation to be correct, haven't
checked it

---------

Co-authored-by: Huy Tran <84343285+huytr1995@users.noreply.github.com>