[OEUI-173] change the login route to always log in regardless of isAuthenticated (#588)
## Description

Closes? OEUI-173

Some users were encountering a login loop. Clearing cookies and tokens
would fix it.

We had a small optimization from our auth0 template that would skip
logging a user in if they had a token. This PR skips that and assumes
that you should get a new token if you're on the `/login` route. This
seems to fix the login loop issue.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Tested locally by tweaking the JWT in localstorage.