[OEUI-144] Create Parameters and Dataset tabs (#597)
## Description

Closes OEUI-144

This PR converts the desktop versions of the attribution and parameters
drawers to be tabs on the thread page. It takes a few liberties with the
formatting of the tab selection buttons to make things easier.

This PR DOES NOT include changes to the contents of the drawers, those
will be coming in other tickets.

Our `primary.dark` seems a little different, but that's a varnish
problem to solve!

## Output / Screenshots

Desktop:

![image](https://github.com/user-attachments/assets/8fb61ade-31ac-4131-afe3-0d37270f1577)

Switching between tabs and scrolling:


https://github.com/user-attachments/assets/2c4f2c63-ba00-4487-9035-2e7bfe7183cf


Just the Dataset tab on desktop:


![image](https://github.com/user-attachments/assets/e3595c62-be38-434b-858d-82bd7e0727d1)

Parameters drawer on mobile:

![image](https://github.com/user-attachments/assets/a765600d-2e6b-4840-8396-9a64b2feaad1)

Attribution drawer on mobile:


![image](https://github.com/user-attachments/assets/6e8ab713-2f2b-4c0d-8824-5f58b5783e97)


### A11y

I've added MUI's suggestions for a11y to the tabs that MUI doesn't add
themselves, including `aria-control` and `aria-labelled-by`.