[OEUI-150], [OEUI-192] FAQ page rebrand (#605)
## Description

Describe the change and its purpose. Include a ticket if applicable.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

Desktop-ish

![image](https://github.com/user-attachments/assets/beed859a-2946-45a7-9901-11dfa01227c3)

Mobile

![image](https://github.com/user-attachments/assets/6747a56f-2910-41ab-a738-c039231f83d7)

![image](https://github.com/user-attachments/assets/91c92654-a9fc-4dd8-b3f8-ac1a5297c716)