[OEUI-237] Add "Open in Dataset Explorer" link to CorpusLink document cards (#637)
## Description

Closes OEUI-237

This PR adds the "Open in Dataset Explorer" button to the CorpusLink
document cards. It links to the corresponding document in the dataset
explorer. The user currently has to click on the button itself, but we
can make it so clicking on any part of the card goes to the right place.

It also makes us fetch 1mm tokens for a document in dataset explorer
instead of just 10. We may want to add a way to fetch all tokens to
infinigram-api at some point, 1mm is just arbitrarily large.

Question: do we want the document to open in another tab?

## Testing Plan

Tested locally. I don't think this needs a test.

## Output / Screenshots


![image](https://github.com/user-attachments/assets/cada2202-261b-43cc-9f46-98a54473906d)

(ignore that the document isn't the same between the document card and
the explorer, this is a mock data thing)

https://github.com/user-attachments/assets/a2e61450-dfd5-4302-99a3-1aa1936fc676

### A11y

The button follows proper a11y practices