[OEUI-233] Dataset Explorer: Colors and styling updated to match figma (#642)
## Description

Figma:
https://www.figma.com/design/lzeziiZpknJSelQF6558d3/AI2-Playground-with-Rebrand-(Sept-2024)?node-id=1262-16349&node-type=frame&t=jYdZkseWKic8XBbJ-0

## Reviewer Instructions

@the-jennajames the reason why the tab that has Search Dataset, Sources,
Web Domains has a white background because when we scroll it will follow
along with it. So without it will look like its overlap so thats why we
have to set a white background for it. Also for Sources the text on the
left i cant align it like your design since it just the limitation of
Nivo Bar chart.
## Testing Plan

Verify UI works. If CI passes we are good.

## Output / Screenshots


https://github.com/user-attachments/assets/def4e94b-5688-40f2-acab-d0c1cf604230

### A11y

N/A