[OEUI-258] Close drawers on navigation (#651)
## Description

The drawers close when the threads are clicked on.

I believe this is the correct behavior. Before this change, the rest of
the UI is grayed out when its open, which doesn't indicate you can do
anything. What's worse is on mobile you might not even notice that the
UI has changed under the drawer and it just seems non responsive.

## Reviewer Instructions

I'm not super happy with the two drawers closing UX, but it is still
better than before. I think this is fine for an iterative approach, and
the whole many drawers situation can be discussed/figured out at a later
time.

## Testing Plan

Manually

## Output / Screenshots


https://github.com/user-attachments/assets/3733f48a-7b0b-4a3c-8cd3-cd185d510379